Fixed Sonar Errors

+++ src/main/java/org/recap/camel/EtlDataLoadProcessor.java
@@ -95,7 +95,7 @@ public class EtlDataLoadProcessor {
         long newItemsCount = itemDetailsRepository.countByOwningInstitutionId(instIdByFileName);
         long newBibHoldingsCount = bibliographicDetailsRepository.findCountOfBibliographicHoldingsByInstId(instIdByFileName);
         long newBibItemsCount = itemDetailsRepository.findCountOfBibliographicItemsByInstId(instIdByFileName);
-        Integer processedBibsCount = (int) (long) newBibsCount - (int) (long) oldBibsCount;
+        Integer processedBibsCount = (int) newBibsCount - (int)oldBibsCount;
         Integer processedHoldingsCount = Integer.valueOf(new Long(newHoldingsCount).toString()) - Integer.valueOf(new Long(oldHoldingsCount).toString());
         Integer processedItemsCount = Integer.valueOf(new Long(newItemsCount).toString()) - Integer.valueOf(new Long(oldItemsCount).toString());
         Integer processedBibHoldingsCount = Integer.valueOf(new Long(newBibHoldingsCount).toString()) - Integer.valueOf(new Long(oldBibHoldingsCount).toString());
